requirements.txt
Flask==2.3.2
flask-cors==3.1.2
google-generativeai==0.1.0
gtts==2.2.3
